added all steps stuff and evrything

+++ src/main/java/com/nighthawk/spring_portfolio/mvc/steptrack1/Day.java
@@ -19,7 +19,7 @@ public class Day {
     private int day;
     private int steps;
     private int calories;
-    private int distanceMiles;
+    private double distanceMiles;
  
     // relationship
     @ManyToOne(fetch = FetchType.LAZY)
